Eliminated IProjectPersistenceReader.GetExistingBooks
Added new static helper class to GlyssenShared to get book IDs in canonical order.
Moved static ForEachBookFileInProject to ProjectRepository.
